add outer data dict

+++ iorder_server/src/main/java/cn/net/normcore/iorder/service/ShopService.java
@@ -28,7 +28,7 @@ public class ShopService extends BaseService {
 						map.put("name", rst.getString("name"));
 						map.put("cheap", rst.getString("cheap"));
 						map.put("picture", rst.getString("picture"));
-						map.put("score", rst.getInt("score"));
+						map.put("score", rst.getFloat("score"));
 						map.put("perPri", rst.getFloat("perPri"));
 						map.put("toSal", rst.getInt("toSal"));
 						map.put("distance", rst.getInt("distance"));
